fix figma colors script to handle our new color doc and generate new colors file (#1747)
## Motivation and Context

This script was not working with our figma layout, and especially with our new CO Design Library. We need the new colors to start building the new theme.

## Description

Fix up the script to handle FRAME as well as RECTANGLE, change how fills are handled, and handle fetching the colors from elements that do have children.